+++ core/xml-reader/src/main/java/org/marceloleite/mercado/xml/adapters/OrderStatusXmlAdapter.java
@@ -4,16 +4,16 @@ import javax.xml.bind.annotation.adapters.XmlAdapter;
 
 import org.marceloleite.mercado.commons.OrderStatus;
 
-public class OrderStatusXmlAdapter extends XmlAdapter<Integer, OrderStatus> {
+public class OrderStatusXmlAdapter extends XmlAdapter<Long, OrderStatus> {
 
 	@Override
-	public OrderStatus unmarshal(Integer orderStatusValue) throws Exception {
+	public OrderStatus unmarshal(Long orderStatusValue) throws Exception {
 		
 		return OrderStatus.getByValue(orderStatusValue);
 	}
 
 	@Override
-	public Integer marshal(OrderStatus orderStatus) throws Exception {
+	public Long marshal(OrderStatus orderStatus) throws Exception {
 		return orderStatus.getValue();
 	}
